fadhil-bithealth Belajar-Training-Bithealth main Project-T1-Final/requirements.txt
langchain
langchain_google_genai
langgraph
pydantic
fastapi
uvicorn
python-multipart
Pillow
python-dotenv
pandas
numpy
qdrant-client
Scikit-learn
pydantic_settings